Code clean up - Phase 1 #3
Refer to the issue mentioned above for more details
