+++ .gitignore
@@ -1,1 +1,3 @@
 *.pro.user*
+.idea
+CMakeLists.txt.user
