+++ core_GiST/gist_compat.h
