+++ core_GiST/gist_cursor.cpp
