+++ core_GiST/gist_cursor.h
