+++ core_GiST/gist_cursorext.cpp
