+++ core_GiST/gist_cursorext.h
