+++ core_GiST/gist_htab.cpp
