+++ core_GiST/gist_p.cpp
