+++ core_GiST/gist_query.cpp
