+++ core_GiST/gist_support.cpp
