+++ core_GiST/vec_t.cpp
