+++ headers/include_handler.hpp
@@ -1,0 +1,12 @@
+#ifndef INCLUDE_HANDLER_HPP
+#define INCLUDE_HANDLER_HPP
+
+
+#include "macros.hpp"
+#include "configs.hpp"
+#include "includes.hpp"
+#include "types.hpp"
+#include "part_class.hpp"
+
+
+#endif // INCLUDE_HANDLER_HPP
