+++ headers/macros.hpp
@@ -1,0 +1,6 @@
+#ifndef MACROS_HPP
+#define MACROS_HPP
+
+#define _GNU_SOURCE 1   //For NOATIME in open(2) //Must be very first line!
+
+#endif // MACROS_HPP
